moved class Vtu into constructive_solid_geometry.py and added set_meta function
